+++ apps/user/src/app/grade-simulation/page.tsx
@@ -1,0 +1,5 @@
+const GradeSimulationPage = () => {
+    return <div>asd</div>;
+};
+
+export default GradeSimulationPage;
