+++ apps/user/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useApiError } from './useApiError';
+export { default as useGradeCalculation } from './useGradeCaculation';
 export { default as useUser } from './useUser';
